Lis algo

+++ stack_init.c
@@ -12,7 +12,7 @@
 
 #include "push_swap.h"
 
-t_node	*new_node(long value)
+t_node	*new_node(int value)
 {
 	t_node	*node;
